move apply guardrails endpoint
